requirement.txt
asgiref==3.5.2
captcha==0.4
Django==3.2.13
django-quill-editor==0.1.40
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-taggit==3.0.0
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
whitenoise==5.3.0
gunicorn==19.1.0